spring day7 commit

+++ test1/src/main/java/com/example/test1/model/Product.java
@@ -5,11 +5,11 @@ import lombok.Data;
 @Data
 public class Product {
 	private String itemNo;
-    private String itemName;
-    private Double price;
-    private Double sRate;
-    private Double pRate;
-    private String contents;
-    private String sellYn;
-    private String transInfo;
+	private String itemName;
+	private int price;
+	private int sRate;
+	private int pRate;
+	private String contents;
+	private String sellYn;
+	private String transInfo;
 }
